split dashboard card components added checked slide-toggle
