+++ src/app/core/dashboard/dashboard-card/dashboard-category-progress-board/dashboard-category-progress-board/dashboard-category-progress-board.component.css
