+++ src/app/core/dashboard/dashboard-card/dashboard-category-progress-board/dashboard-category-progress-board/dashboard-category-progress-board.component.ts
@@ -1,5 +1,5 @@
 import {Component, Input, OnInit} from '@angular/core';
-import {DashboardProgressBar} from '../dashboard-progress-bar/dashboard-progress-bar.model';
+import {DashboardProgressBar} from '../../dashboard-progress-bar/dashboard-progress-bar.model';
 
 @Component({
   selector: 'app-dashboard-category-progress-board',
