+++ src/app/core/dashboard/dashboard-card/dashboard-category-progress-board/dashboard-category-share-board/dashboard-category-share-board.component.ts
@@ -4,7 +4,7 @@ import {DashboardProgressBar} from '../../dashboard-progress-bar/dashboard-progr
 @Component({
   selector: 'app-dashboard-category-share-board',
   templateUrl: './dashboard-category-share-board.component.html',
-  styleUrls: ['../dashboard-category-progress-board.component.css']
+  styleUrls: ['../dashboard-category-progress-board/dashboard-category-progress-board.component.css']
 })
 export class DashboardCategoryShareBoardComponent implements OnInit {
   @Input() categoryA: string;
