+++ src/app/core/dashboard/dashboard-card/dashboard-progress-radial-bar/dashboard-progress-radial-bar.component.css
@@ -1,0 +1,4 @@
+#chart {
+  max-width: 350px;
+  margin-bottom: -80px !important;
+}
