+++ src/app/core/dashboard/dashboard-card/dashboard-progress-radial-bar/dashboard-progress-radial-bar.component.html
@@ -1,0 +1,11 @@
+<div id="chart">
+  <apx-chart
+    [chart]="indicatorChartOptions.chart"
+    [fill]="indicatorChartOptions.fill"
+    [labels]="indicatorChartOptions.labels"
+    [plotOptions]="indicatorChartOptions.plotOptions"
+    [series]="indicatorChartOptions.series"
+    [stroke]="indicatorChartOptions.stroke"
+    [theme]="indicatorChartOptions.theme"
+  ></apx-chart>
+</div>
