+++ src/app/core/dashboard/dashboard-card/dashboard-rank/dashboard-rank.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DashboardRankComponent } from './dashboard-rank.component';
+
+describe('DashboardRankComponent', () => {
+  let component: DashboardRankComponent;
+  let fixture: ComponentFixture<DashboardRankComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ DashboardRankComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DashboardRankComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
