+++ src/app/core/dashboard/ppr-dashboard/ppr-dashboard.component.html
@@ -4,5 +4,4 @@
   [userCode]="userCode"
   [radialBarSeries]="radialBarSeries"
   [employees]="employees"
-  test="Oleh"
 ></app-dashboard-card>
